delete és update függvényekben hibák lekezelése

+++ app/Http/Controllers/API/CarController.php
@@ -46,6 +46,9 @@ class CarController extends Controller
     public function update(Request $request, string $id)
     {
         $car = Car::find($id);
+        if(is_null($car)){
+            return response()->json(["message" => "No item found with id: $id"], 404);
+        }
         $car->fill($request->all());
         $car->save();
         return $car;
@@ -57,6 +60,10 @@ class CarController extends Controller
     public function destroy(string $id)
     {
         $car = Car::find($id);
+        if(is_null($car)){
+            return response()->json(["message" => "No item found with id: $id"], 404);
+        }
         $car->delete();
+        return response()->noContent();
     }
 }
